feat: add useCase of updateBoardUserPhase
