+++ backend/src/modules/boards/boards.module.ts
@@ -36,8 +36,8 @@ import {
 	sendBoardTimerTimeLeftService,
 	startBoardTimerService,
 	stopBoardTimerService,
-	updateBoardApplication,
 	updateBoardParticipantsUseCase,
+	updateBoardPhaseUseCase,
 	updateBoardService,
 	updateBoardTimerDurationService,
 	updateBoardUseCase
@@ -76,8 +76,8 @@ import { VotesModule } from '../votes/votes.module';
 		isBoardPublicUseCase,
 		updateBoardUseCase,
 		updateBoardParticipantsUseCase,
+		updateBoardPhaseUseCase,
 		mergeBoardUseCase,
-		updateBoardApplication,
 		deleteBoardUseCase,
 		boardTimerRepository,
 		sendBoardTimerStateService,
