+++ backend/src/modules/boards/boards.providers.ts
@@ -18,7 +18,6 @@ import { GetBoardUseCase } from './applications/get-board.use-case';
 import { GetBoardsForDashboardUseCase } from './applications/get-boards-for-dashboard.use-case';
 import { GetPersonalBoardsUseCase } from './applications/get-personal-boards.use-case';
 import { IsBoardPublicUseCase } from './applications/is-board-public.use-case';
-import { UpdateBoardApplication } from './applications/update.board.application';
 import { TYPES } from './interfaces/types';
 import { BoardRepository } from './repositories/board.repository';
 import CreateBoardService from './services/create.board.service';
@@ -28,6 +27,7 @@ import UpdateBoardService from './services/update.board.service';
 import { UpdateBoardUseCase } from './applications/update-board.use-case';
 import { UpdateBoardParticipantsUseCase } from './applications/update-board-participants.use-case';
 import { MergeBoardUseCase } from './applications/merge-board.use-case';
+import { UpdateBoardPhaseUseCase } from './applications/update-board-phase.use-case';
 
 export const createBoardService = {
 	provide: TYPES.services.CreateBoardService,
@@ -94,9 +94,9 @@ export const mergeBoardUseCase = {
 	useClass: MergeBoardUseCase
 };
 
-export const updateBoardApplication = {
-	provide: TYPES.applications.UpdateBoardApplication,
-	useClass: UpdateBoardApplication
+export const updateBoardPhaseUseCase = {
+	provide: TYPES.applications.UpdateBoardPhaseUseCase,
+	useClass: UpdateBoardPhaseUseCase
 };
 
 export const deleteBoardUseCase = {
