--- backend/src/modules/boards/interfaces/applications/update.board.application.interface.ts
@@ -1,5 +1,0 @@
-import { BoardPhaseDto } from 'src/libs/dto/board-phase.dto';
-
-export interface UpdateBoardApplicationInterface {
-	updatePhase(payload: BoardPhaseDto);
-}
