+++ backend/src/modules/boards/interfaces/services/update.board.service.interface.ts
@@ -1,10 +1,5 @@
 import { TeamDto } from 'src/modules/communication/dto/team.dto';
-import { BoardPhaseDto } from 'src/libs/dto/board-phase.dto';
 
 export interface UpdateBoardServiceInterface {
-	//mergeBoards(subBoardId: string, userId: string, socketId?: string): Promise<Board> | null;
-
 	updateChannelId(teams: TeamDto[]);
-
-	updatePhase(payload: BoardPhaseDto);
 }
